+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # Unreleased
 
+* Add support for converting 1536 well plates in cellvoyager converters(\#715).
 * Make validation of NGFF Zarr attributes more strict, in tests (\#732).
 
 # 1.0.0